requirements.txt
ffmpeg-python==0.2.0
elevenlabs==0.2.27
replicate==0.22.0
